Constexpr scalars don't typecast in binary ops when they have the a type of comparable kind as the tensor
The idea here is that if you have a tensor `t` of dtype `uint8` and you want
to do `t << 2`, the result should be of dtype `uint8`, not `int32`!

We do this for all dunder ops that don't output booleans.

This follows roughly the semantics of PyTorch, JAX and NumPy 2.0.

I would like to document this behaviour, but it's not clear to me where
is the best place to say so.

The PR has much more churn than I would like, as I had to move the
`to_tensor` method to `semantic` (which is where it belongs anyway).
For reviewers, the only two relevant changes are in `computation_type_impl` and
in `bitwise_op_type_checking_impl`, where we say that we do perform casting
for bitwise ops.